[refactor] di factory 형식 통일
