+++ FLINT/FLINT/Dependency/Factory/Service/HomeServiceFactory.swift
@@ -1,0 +1,26 @@
+//
+//  HomeServiceFactory.swift
+//  FLINT
+//
+//  Created by 김호성 on 2026.01.31.
+//
+
+import Foundation
+
+import Moya
+
+import Data
+
+protocol HomeServiceFactory: HomeAPIFactory {
+    func makeHomeService() -> HomeService
+    func makeHomeService(homeAPIProvider: MoyaProvider<HomeAPI>) -> HomeService
+}
+
+extension HomeServiceFactory {
+    func makeHomeService() -> HomeService {
+        return makeHomeService(homeAPIProvider: makeHomeAPIProvider())
+    }
+    func makeHomeService(homeAPIProvider: MoyaProvider<HomeAPI>) -> HomeService {
+        return DefaultHomeService(provider: homeAPIProvider)
+    }
+}
