+++ FLINT/FLINT/Dependency/Factory/ViewController/CreateCollectionViewControllerFactory+.swift
@@ -9,7 +9,7 @@ import Foundation
 
 import Presentation
 
-extension CreateCollectionViewControllerFactory where Self: CreateCollectionFactory & ViewControllerFactory {
+extension CreateCollectionViewControllerFactory where Self: CreateCollectionViewModelFactory & ViewControllerFactory {
     func makeCreateCollectionViewController() -> CreateCollectionViewController {
         let vm = makeCreateCollectionViewModel()
         return CreateCollectionViewController(viewModel: vm, viewControllerFactory: self)
