--- FLINT/FLINT/Dependency/Factory/ViewModel/CollectionFolderListViewModel.swift
@@ -1,7 +1,0 @@
-//
-//  CollectionFolderListViewModel.swift
-//  FLINT
-//
-//  Created by 소은 on 1/24/26.
-//
-
